// src/types/GpsMeta.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type GpsMeta = { lat: number, lng: number, alt: number | null, altRef: string | null, direction: number | null, directionRef: string | null, speed: number | null, timestamp: string | null, processingMethod: string | null, dop: number | null, };